Add EncapsulationKind enum, big-endian uint reader/writers
The second byte of the CDR header is an EncapsulationKind enum, not a simple littleEndian boolean. This PR adds support for reading CdrReader#kind and changes the constructor options for CdrWriter to take an optional kind arg instead of bigEndian.

The other change is four new methods on CdrReader and CdrWriter each for reading and writing unsigned integers as big-endian regardless of the endianness of the reader/writer. This is useful for bitflags, enum values, and guids.