Fix alignment for empty typed arrays (#7)
Sequences have no special alignment requirements, which means that only the individual primitive values in the sequence need to be properly aligned. This means that an empty array has no alignment requirements other than the uint32 sequence length before it. As such, the `this.align()` call is incorrect if the array length is 0.